Move Message to Dialogs
