+++ UI/Dialogs/Message.xaml
@@ -1,5 +1,5 @@
 ﻿<ContentDialog
-    x:Class="SchemeCreator.UI.Message"
+    x:Class="SchemeCreator.UI.Dialogs.Message"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:SchemeCreator.UI"
