+++ UI/Dialogs/Message.xaml.cs
@@ -1,7 +1,7 @@
 ﻿using SchemeCreator.Data.Services;
 using Windows.UI.Xaml.Controls;
 
-namespace SchemeCreator.UI
+namespace SchemeCreator.UI.Dialogs
 {
     public sealed partial class Message : ContentDialog
     {
